Removed 'Leaders' group.
This is no longer needed or really used on the site, since the logic
accounts for Users who are leaders by virtue of being a Camp leader.
Security is therefore improved because only leaders of current camps are
given access, and the Leaders group does not need to be kept up to date.

This also involved fixing various tests, including test isolation issues and
mocking get_thisyear().